+++ app/assets/config/grape_swagger_rails_manifest.js
@@ -1,0 +1,2 @@
+//= link_directory ../stylesheets/grape_swagger_rails .css
+//= link_directory ../javascripts/grape_swagger_rails .js