+++ spec/dummy/app/assets/javascripts/application.js
@@ -1,3 +1,1 @@
-//= require jquery
-//= require jquery_ujs
 //= require_tree .