+++ spec/dummy/config/boot.rb
@@ -1,10 +1,5 @@
-require 'rubygems'
-
-ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../../../Gemfile', __FILE__)
+# Set up gems listed in the Gemfile.
+ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../../Gemfile', __dir__)
 
 require 'bundler/setup' if File.exist?(ENV['BUNDLE_GEMFILE'])
-
-require 'sass'
-require 'grape'
-require 'grape-swagger'
-require 'coffee_script'
+$LOAD_PATH.unshift File.expand_path('../../../lib', __dir__)