+++ spec/dummy/config/environment.rb
@@ -1,3 +1,5 @@
-require File.expand_path('../application', __FILE__)
+# Load the Rails application.
+require_relative 'application'
 
-Dummy::Application.initialize!
+# Initialize the Rails application.
+Rails.application.initialize!