+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Next Release
 
 * Your contribution here.
+* [#110](https://github.com/ruby-grape/grape-swagger-rails/pull/110): Update dummy app to current rails conventions - [@duffn](https://github.com/duffn).
 
 ### 0.4.0 (2023/03/28)